refactor(model): extract aggregation builders
